Implementation of Builder Design Pattern in doctoshotgun.py
I applied the builder pattern on the lines from #630 to #669 of the original code, in “main” function of “Application” class, where the parser is repeatedly called and adds arguments to itself. I imported “ABC” and “abstractmethod” from the “abc” to make the interface for the builder I wanted to have. The interface name is “IArgumentBuilder”, and it has the abstract methods of some of the functions in the builder. Then I made my builder class “ArgumentBuilder” which is my concrete builder class where all the functions make some part of the parser. I got the names of the functions in my builder class from what they do so it would be easier for the reader to understand what part of the object is being constructed in that specific function. Then I have my “Director” class which constructs the object in the sequence that I defined and can be changed easily if there are changes must be made.